capmonster/local/requirements.txt
python-dotenv~=0.19.1
certifi~=2021.5.30
motor~=2.5.1
pymongo~=3.12.1
pydantic~=1.8.2
httpx~=0.20.0